pyproject.toml
[tool.poetry]
name = "blaise-totalmobile-client"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
blaise-restapi = {git = "https://github.com/ONSdigital/blaise-api-python-client.git", rev = "main"}
google-auth = "^2.6.0"
google-cloud-pubsub = "^2.11.0"
google-cloud-tasks = "^2.8.1"
Flask = "^2.0.3"
config = "^0.5.1"
python-dotenv = "^0.19.2"
pytest = "^7.1.1"
Flask-HTTPAuth = "^4.5.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pytest = "^7.0.1"
python-dotenv = "^0.19.2"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.931"
types-requests = "^2.27.11"
types-Flask = "^1.1.6"
pytest-cov = "^3.0.0"
requests-mock = "^1.9.0"
Flask = "^2.0.3"
Flask-SQLAlchemy = "^2.5.1"
SQLAlchemy = "^1.4.32"
phonenumbers = "^8.12.46"
pytest-lazy-fixture = "^0.6.3"
mocker = "^1.1.1"
responses = "^0.20.0"
requests-mock-flask = "^2022.4.3"

[tool.isort]
multi_line_output = "3"
include_trailing_comma = "True"
line_length = "88"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
